Add and use SecondaryExpandablePanel
