--- components/articles/article-alinea-0/index.js
@@ -1,3 +1,0 @@
-import ArticleAlinea0 from "./article-alinea-0-component";
-
-export default ArticleAlinea0;
