+++ components/articles/articles-component.jsx
@@ -9,7 +9,6 @@ import {
 } from "../articles-inputs";
 import { Parameter } from "../articles-inputs/parameter";
 import fillArrayWith from "../utils/array/fillArrayWith";
-import Alinea0 from "./article-alinea-0";
 import Alinea2 from "./article-alinea-2";
 import Alinea3 from "./article-alinea-3";
 import Alinea4a from "./article-alinea-4a";
@@ -17,7 +16,7 @@ import ArticleHeader from "./article-header";
 import BoutonAjouterTranche from "./article-tranches/bouton-ajouter-tranche";
 import BoutonSupprimerTranche from "./article-tranches/bouton-supprimer-tranche";
 import styles2 from "./articles-component.module.scss";
-import { PrimaryParameterExpandablePanel as PrimaryExpandablePanel } from "./expandable-panels/PrimaryParameterExpandablePanel";
+import { PrimaryParameterExpandablePanel as PrimaryExpandablePanel, SecondaryExpandablePanel } from "./expandable-panels";
 
 const style = {
   Button: {
@@ -186,7 +185,10 @@ class ArticlesComponent extends React.Component {
             subTitle="Article 197 du CGI - I.1"
             title="Barème et taux"
           >
-            <Alinea0 style={style} />
+            <SecondaryExpandablePanel title="I. En ce qui concerne les contribuables ...">
+              visés à l&apos;article 4 B, il est fait application des règles
+              suivantes pour le calcul de l&apos;impôt sur le revenu :
+            </SecondaryExpandablePanel>
             {articleTranches}
             <Grid container spacing={0}>
               <Grid item sm={3}>
