+++ components/articles/expandable-panels/SecondaryExpandablePanel.module.scss
@@ -1,0 +1,10 @@
+@import 'styles/variables';
+
+.title {
+  font-family: $law-font-family;
+  margin-left: -20px;
+}
+
+.content {
+  font-family: $law-font-family;
+}
