+++ components/articles/expandable-panels/index.js
@@ -1,0 +1,3 @@
+export { PrimaryExpandablePanel } from "./PrimaryExpandablePanel";
+export { PrimaryParameterExpandablePanel } from "./PrimaryParameterExpandablePanel";
+export { SecondaryExpandablePanel } from "./SecondaryExpandablePanel";
